% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt_districts}
\alias{bt_districts}
\title{District information}
\format{A data frame with nine variables:}
\source{
\url{http://www.bundeswahlleiter.de/en/}
}
\usage{
data(bt_districts)
}
\description{
Demographic and social information on Bundestag districts
}
\keyword{datasets}

