% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt_districts_maps}
\alias{bt_districts_maps}
\title{District Maps}
\format{A data frame with 6 variables:
\describe{
\item{\code{year}}{the election year}
\item{\code{district}}{the district name}
\item{\code{wk}}{the district number}
\item{\code{long}}{longitudional info}
\item{\code{lat}}{latidutional info}
}}
\usage{
data(bt_districts_maps)
}
\description{
Maps of Bundestag districts
}
\details{
Polygon format
}
\keyword{datasets}

