% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt_districts_neighbors}
\alias{bt_districts_neighbors}
\title{Adjacent districts}
\format{A data frame:
\describe{
\item{\code{id}}{number for district 1}
\item{\code{id_from}}{number for district 2}
\item{\code{dis}}{distance between \code{id} and \code{id_from}}
\item{\code{year}}{year of the election}
\item{\code{wp}}{the legislative period}
}}
\usage{
data(bt_districts_neighbors)
}
\description{
Closeness of Bundestag districts
}
\keyword{datasets}

