% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btw_candidates}
\alias{btw_candidates}
\title{Candidates.}
\format{A data frame with 18 variables:
\describe{
\item{\code{year}}{year of the election}
\item{\code{wp}}{numeric information on the legislative period}
\item{\code{bundesland}}{the Bundesland in which the candidate applied}
\item{\code{wk}}{the district number}
\item{\code{party}}{the party name}
\item{\code{lname}}{candidate's last name}
\item{\code{fname}}{candidate's first name}
\item{\code{age}}{candidate's age}
\item{\code{gender}}{candidate's gender}
\item{\code{dr}}{does candidate hold a PhD?}
\item{\code{adel}}{does candidate have a noble name?}
\item{\code{german}}{does candidate have a german name?}
\item{\code{on_list}}{is candidate' name on the party list?}
\item{\code{on_wk}}{does the candidate run in a district?}
\item{\code{toplevel}}{did candidate obtain high-level political office?}
\item{\code{mdl}}{was candidate member of a state parliament before?}
\item{\code{mdb}}{was candidate member of the Bundestag before?}
}}
\source{
\url{http://www.bundeswahlleiter.de/en/} + self compiled from various sources
}
\usage{
data(btw_candidates)
}
\description{
Information on candidates in Bundestag elections. Currently, information are
only available from 1980 onward.
}
\keyword{datasets}

