% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btw_districts}
\alias{btw_districts}
\title{District information.}
\format{An object of class \code{data.frame} with 4904 rows and 68 columns.}
\source{
\url{http://www.bundeswahlleiter.de/en/}
}
\usage{
data(btw_districts)
}
\description{
Demographic and social information on Bundestag districts. 
Mostly from 1980 onwards.
}
\keyword{datasets}

