% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btw_districts_maps}
\alias{btw_districts_maps}
\title{District Maps.}
\format{A data frame with seven variables:
\describe{
\item{\code{year}}{the election year}
\item{\code{bundesland}}{Bundesland of the district}
\item{\code{district}}{the district name}
\item{\code{wk}}{the district number}
\item{\code{part}}{running index for subareas of districts (e.g. islands)}
\item{\code{long}}{longitudional info}
\item{\code{lat}}{latidutional info}
}}
\usage{
data(btw_districts_maps)
}
\description{
Maps of Bundestag districts
}
\details{
Polygon format
}
\keyword{datasets}

