% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btw_votes}
\alias{btw_votes}
\title{Votes.}
\format{A data frame with 47840 rows and seven variables: 
\describe{
\item{\code{year}}{year of the election}
\item{\code{bundesland}}{the Bundesland of the \code{district}}
\item{\code{district}}{the district name}
\item{\code{wk}}{the district number}
\item{\code{party}}{the party name}
\item{\code{mandate}}{type of mandate, cvote = Candidate Vote, pvote = Party Vote}
\item{\code{per}}{the percentage of votes for the \code{party}, NA means party did not 
run a candidate or a party list}
}}
\source{
\url{http://www.bundeswahlleiter.de/en/}
}
\usage{
data(btw_votes)
}
\description{
Complete data on German Bundestag elections form 1949 to 2013
}
\details{
In the 1949 election both tiers were merged into a single vote. In the dataset this vote is registered as the party vote
In 1957 CDU and CSU ran separately in newly annexed Saarland. Here, CSU receives as CSU/CVP label.
}
\keyword{datasets}

