% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btw_districts}
\alias{btw_districts}
\title{District information.}
\format{A data frame with nine variables:
\describe{
\item{\code{year}}{the election year}
\item{\code{bundesland}}{Bundesland of the district}
\item{\code{district}}{the district name}
\item{\code{wk}}{the district number}
\item{\code{part}}{running index for subareas of districts (e.g. islands)}
\item{\code{long}}{longitudional info}
\item{\code{lat}}{latidutional info}
}}
\source{
\url{http://www.bundeswahlleiter.de/en/}
}
\usage{
data(btw_districts)
}
\description{
Demographic and social information on Bundestag districts
}
\keyword{datasets}

